Moved `RolesGuard` into the auth module root
