+++ anthill_back/src/modules/auth/roles.guard.ts
@@ -1,8 +1,8 @@
 import { Observable } from 'rxjs';
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Request } from 'express';
-import { JwtPayloadDto } from '../data/dtos/jwt.payload.dto';
-import { UserRole } from '../../users/data/entities/user.entity';
+import { JwtPayloadDto } from './data/dtos/jwt.payload.dto';
+import { UserRole } from '../users/data/entities/user.entity';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
